Dynamically choose Ingress proxy store
Without this change, the norman store proxy setup chooses the Ingress
API to register based on the local cluster context at startup. It does
not know about other clusters at this point and can't register different
APIs on a cluster-by-cluster basis. This causes a problem when the local
cluster is version 1.18 (does not support networking.k8s.io/v1/Ingress)
and the downstream cluster is 1.22 (does not support
extensions/v1beta1/Ingress) or vice versa, where when the user tries to
navigate to the downstream cluster the wrong Ingress API will have been
registered and the Watch will fail with a 404 error.

This change addresses the problem by first statically registering the
networking.k8s.io/v1/Ingress API into the schema store on the assumption
that this will be used most of the time, and then adding dynamic store
recreation in the Store methods which can detect the cluster that's
actually in use when the request is being made.

This will probably incur some performance overhead for kubernetes 1.18
clusters. These changes should be removed once support for kubernetes
1.18 is dropped.